+++ web/diag/show_realmcheck.php
@@ -34,7 +34,7 @@ $gui = new \web\lib\user\Gui();
 
 $gui->languageInstance->setTextDomain("diagnostics");
 
-echo $deco->defaultPagePrelude(sprintf(_("Sanity check for dynamic discovery of realms"), \config\Master::APPEARANCE['productname']), false);
+echo $deco->defaultPagePrelude(sprintf(_("Sanity check for dynamic discovery of realms"), \config\Main::APPEARANCE['productname']), false);
 
 $ourlocale = $gui->languageInstance->getLang();