+++ web/lib/admin/PageDecoration.php
@@ -99,7 +99,7 @@ class PageDecoration extends \core\common\Entity {
                     <form action='$place' method='GET' accept-charset='UTF-8'>" . _("View this page in") . "&nbsp;
                         <select id='lang' name='lang' onchange='this.form.submit()'>";
 
-        foreach (\config\Master::LANGUAGES as $lang => $getValue) {
+        foreach (\config\Main::LANGUAGES as $lang => $getValue) {
             $retval .= "<option value='$lang' " . (strtoupper($language) == strtoupper($lang) ? "selected" : "" ) . " >" . $getValue['display'] . "</option> ";
         }
         $retval .= "</select>";
@@ -154,7 +154,7 @@ class PageDecoration extends \core\common\Entity {
         // highlighting to work
         $retval = "<div class='maincontent'>";
         // default values which match almost every case; override where needed
-        $cap1 = \config\Master::APPEARANCE['productname_long'];
+        $cap1 = \config\Main::APPEARANCE['productname_long'];
         $advancedControls = TRUE;
         switch ($area) {
             case "ADMIN-PARTICIPANT":
@@ -179,8 +179,8 @@ class PageDecoration extends \core\common\Entity {
                 $cap2 = sprintf(_("Administrator Interface - %s Management"), $this->ui->nomenclatureFed);
                 break;
             case "USER":
-                $cap1 = sprintf(_("Welcome to %s"), \config\Master::APPEARANCE['productname']);
-                $cap2 = \config\Master::APPEARANCE['productname_long'];
+                $cap1 = sprintf(_("Welcome to %s"), \config\Main::APPEARANCE['productname']);
+                $cap2 = \config\Main::APPEARANCE['productname_long'];
                 $advancedControls = FALSE;
                 break;
             case "SUPERADMIN":
@@ -203,9 +203,9 @@ class PageDecoration extends \core\common\Entity {
                 <h2 style='padding-left:10px;'>$cap2</h2>
             </div><!--secondarycaptions-->";
 
-        if (isset(\config\Master::APPEARANCE['MOTD']) && \config\Master::APPEARANCE['MOTD'] != "") {
+        if (isset(\config\Main::APPEARANCE['MOTD']) && \config\Main::APPEARANCE['MOTD'] != "") {
             $retval .= "<div id='header_MOTD' style='display:inline-block; padding-left:20px;vertical-align:top;'>
-              <p class='MOTD'>" . \config\Master::APPEARANCE['MOTD'] . "</p>
+              <p class='MOTD'>" . \config\Main::APPEARANCE['MOTD'] . "</p>
               </div><!--header_MOTD-->";
         }
         $retval .= $this->sidebar($advancedControls);
@@ -270,8 +270,8 @@ class PageDecoration extends \core\common\Entity {
                 <tr>
                     <td style='padding-left:20px; padding-right:20px; text-align:left; vertical-align:top;'>
                         " . $cat->catCopyright . "</td>";
-        if (!empty(\config\Master::APPEARANCE['privacy_notice_url'])) {
-            $retval .= "<td><a href='".\config\Master::APPEARANCE['privacy_notice_url']."'>" . sprintf(_("%s Privacy Notice"),\config\ConfAssistant::CONSORTIUM['display_name']) . "</a></td>";
+        if (!empty(\config\Main::APPEARANCE['privacy_notice_url'])) {
+            $retval .= "<td><a href='".\config\Main::APPEARANCE['privacy_notice_url']."'>" . sprintf(_("%s Privacy Notice"),\config\ConfAssistant::CONSORTIUM['display_name']) . "</a></td>";
         }
         $retval .= "            <td style='padding-left:80px; padding-right:20px; text-align:right; vertical-align:top;'>";