+++ web/lib/user/Gui.php
@@ -42,7 +42,7 @@ class Gui extends \core\UserAPI {
             $fed = $validator->existingFederation($idp->federation);
             $fedskin = $fed->getAttributes("fed:desired_skin");
         }
-        $this->skinObject = new \web\lib\user\Skinjob($_REQUEST['skin'] ?? $_SESSION['skin'] ?? $fedskin[0] ?? \config\Master::APPEARANCE['skins'][0]);
+        $this->skinObject = new \web\lib\user\Skinjob($_REQUEST['skin'] ?? $_SESSION['skin'] ?? $fedskin[0] ?? \config\Main::APPEARANCE['skins'][0]);
         $this->languageInstance->setTextDomain("web_user");
         $this->textTemplates = new TextTemplates();
         $this->operatingSystem = $this->detectOS();
@@ -55,7 +55,7 @@ class Gui extends \core\UserAPI {
      * @param string $pagetitle content for the <title> element
      * @return void writes directly to output
      */
-    public function defaultPagePrelude($pagetitle = \config\Master::APPEARANCE['productname_long']) {
+    public function defaultPagePrelude($pagetitle = \config\Main::APPEARANCE['productname_long']) {
         $ourlocale = $this->languageInstance->getLang();
         header("Content-Type:text/html;charset=utf-8");
         echo "<!DOCTYPE html>